back to config format
