+++ cg/meta/workflow/nf_analysis.py
@@ -152,7 +152,7 @@ class NfAnalysisAPI(AnalysisAPI):
         return Path(self.get_case_path(case_id), NFX_WORK_DIR)
 
     def set_cluster_options(self, case_id: str) -> str:
-        return f'process.clusterOptions: "-A {self.account} --qos={self.get_slurm_qos_for_case(case_id=case_id)}"\n'
+        return f'process.clusterOptions = "-A {self.account} --qos={self.get_slurm_qos_for_case(case_id=case_id)}"\n'
 
     @staticmethod
     def extract_read_files(
